# lib/api-zod/src/generated/types/createJobInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WooCommerce Import Pipeline API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateJobInputConfig } from "./createJobInputConfig";
import type { CreateJobInputType } from "./createJobInputType";

export interface CreateJobInput {
  type: CreateJobInputType;
  storeId?: number | null;
  config?: CreateJobInputConfig;
}