# lib/api-zod/src/generated/types/createStoreInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WooCommerce Import Pipeline API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateStoreInput {
  name: string;
  url: string;
  consumerKey: string;
  consumerSecret: string;
}