# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WooCommerce Import Pipeline API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductRawData } from "./productRawData";
import type { ProductStatus } from "./productStatus";

export interface Product {
  id: number;
  sunskyId: string;
  sku: string;
  name: string;
  description?: string | null;
  price?: string | null;
  stockStatus?: string | null;
  status: ProductStatus;
  categoryId?: string | null;
  imageCount: number;
  wooProductId?: number | null;
  errorMessage?: string | null;
  rawData?: ProductRawData;
  createdAt: string;
  updatedAt: string;
}