# lib/api-zod/src/generated/types/wooCategory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WooCommerce Import Pipeline API
 * OpenAPI spec version: 0.1.0
 */

export interface WooCategory {
  id: number;
  wooId: number;
  name: string;
  slug: string;
  parentId?: number | null;
  count: number;
}